Big Plugin changes. - plugins no longer seperated between api / template. plugins are now just plugins, with two methods   HandleRequest and HandleVirtualRequest. The first method handles a request for a physical resorce on the webserver (ex html file).   The second handles a request for a non existent file on the server (ex an api function call). - moved cache out of main project directoy and in to its own library. This is necessary to allow all plugins to share the cache with the   main web server and with other plugins. Thus if multiple plugins use the same resorce it will be cached across all plugins. - new pluginUtil library that contains handy functions for writing plugins, like parsing and executing a go template file (gohtml) or loading the contents of a file (with cache lookup).
This update is a big shift in thinking form how I intended plugins to work. Plugins now have much more control over how an http request is served.
The server is just there to pass relevant information over to the plugin and it is the plugins job to actually produce the http response. This is apposed to how it used to work
where by plugins hand narrow roles such as, template plugin. In said old system a template plugin would only provide the server with a template parsing / execution strcture and nothing more,
it wouldent even get invoked if the file did not end in .gohtml. In short plugin responsability++ server responsability--.